eleve_parcours_temoignage

+++ src/main/java/net/ubilife/spring/customerjdbc/UserRepositoryImpl.java
@@ -19,7 +19,7 @@ public class UserRepositoryImpl implements UserRepository {
 
 	@Autowired
 	private JdbcOperations jdbc;
-	
+
 	private static final String SQL_INSERT = "insert into user (login, password, nom, nomFamille, prenom, type, numero, mail, IDPARCOURS) values (?,?,?,?,?,?,?,?,?)";
 	private static final String SQL_UPDATE = "update user set login=?, password=?, nom=?, nomFamille=?, prenom=?, type=?, numero=?, mail=?, IDPARCOURS=?";
 	private static final String SQL_UPDATE_ONE = "update user set login=?, password=?, nom=?, nomFamille=?, prenom=?, type=?, numero=?, mail=?, IDPARCOURS=? where userId=?";
@@ -34,20 +34,20 @@ public class UserRepositoryImpl implements UserRepository {
 	public User findOne(long id) {
 		return jdbc.queryForObject(SQL_FIND_ONE, new UserRowMapper(), id);
 	}
-	
+
 
 
 	@Override
 	public User save(final User user) {
-		
+
 		KeyHolder holder = new GeneratedKeyHolder();
-		
+
 		int rows = jdbc.update(new PreparedStatementCreator() {
-			
+
 			@Override
 			public PreparedStatement createPreparedStatement(Connection conn) throws SQLException {
 				PreparedStatement ps = conn.prepareStatement(SQL_INSERT, new String[]{"userId"});
-				
+
 				ps.setString(1, user.getLogin());
 				ps.setString(2, user.getPassword());
 				ps.setString(3, user.getNom());
@@ -60,14 +60,14 @@ public class UserRepositoryImpl implements UserRepository {
 				return ps;
 			}
 		}, holder);
-		
+
 		if(rows == 1) {	// success, so apply ID to the customer object
-			user.setId((Long)holder.getKey());
+			user.setId((Integer)holder.getKey());
 			return user;
 		}
-		
+
 		return null;
-		
+
 	}
 
 	@Override
@@ -105,10 +105,10 @@ public class UserRepositoryImpl implements UserRepository {
 
 		@Override
 		public User mapRow(ResultSet rs, int row) throws SQLException {
-			
+
 			return new User(rs.getInt("userId"), rs.getString("login"), rs.getString("password"), rs.getString("nom"), rs.getString("nomFamille"), rs.getString("prenom"), rs.getString("type"), rs.getString("numero"), rs.getString("mail"),rs.getInt("IDPARCOURS"));
-			
+
 		}
-		
+
 	}
 }
